requirements.txt
# Core dependencies
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.66.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# API client dependencies
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
tenacity>=8.2.0  # for retry logic

# Google Cloud dependencies for Veo
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0  # For GCS video storage
google-auth>=2.23.0
google-auth-httplib2>=0.1.1

# Video processing
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
moviepy>=1.0.3

# Evaluation metrics
scikit-image>=0.22.0
scipy>=1.11.0
torch>=2.0.0  # for perceptual metrics
torchvision>=0.15.0

# Data handling
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Logging and monitoring
loguru>=0.7.2
rich>=13.7.0  # for beautiful terminal output

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.7.0
types-requests>=2.31.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
